pyproject.toml
[project]
name = "sample-python-project"
version = "0.0.1"
dependencies = [
    "python-dotenv"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sample_python_project"]

[project.scripts]
my-cli = "sample_python_project.main:cli"

[tool.black]
line-length = 100

[tool.ruff]
# Never enforce `E501` (line length violations).
ignore = ["C901", "E501", "E741"]
select = ["C", "E", "F", "W"]
line-length = 100
# Ignore import violations in all `__init__.py` files.
[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.isort]
profile = "black"
line_length = 100
